requirements.txt
Pillow
python-csv
